requirements.txt
pytubefix>=3.5.0
moviepy>=1.0.3
openai-whisper>=20231117
colorama>=0.4.6
python-dotenv>=1.0.0
torch>=2.1.0
agno>=0.1.0 